Bugs

+++ src/main/java/kristiania/no/http/controllers/AddAndListAllQuestionsController.java
@@ -39,7 +39,7 @@ public class AddAndListAllQuestionsController implements HttpController {
 
             responseText = "You have added: Question: " + q.getTitle() + "\r\n Survey: " + q.getSurveyId() + "\r\nWith options:";
             for (int i = 1; i < 6; i++) {
-                Option option = new Option(queryMap.get("option" + i), (int) q.getId());
+                Option option = new Option(queryMap.get("option" + i), q.getId());
                 if (option.getOptionName() != "") {
                     optionDao.save(option);
                 }
